refactor(tools): name files after their tool directory for consistency
Three tools had files that didn't follow the repo convention of
filename == tool directory name:
- json-to-yaml-converter: json-to-yaml.* -> json-to-yaml-converter.*
- yaml-to-json-converter: yaml-to-json.* -> yaml-to-json-converter.*
- otp-code-generator-and-validator: otp.service[.test].ts and
  otp-code-generator.e2e.spec.ts -> otp-code-generator-and-validator.*

Rename each tool's component/service/test/e2e files to match its
directory and update the local imports (component lazy-import in
index.ts, service imports in the .vue and .service.test). Pure rename +
import path updates — no logic changes; token-display.vue (a genuine
sub-component) is left as-is. Visual-regression goldens are keyed by
directory name, so they are unaffected.
