+++ src/tools/json-to-yaml-converter/index.ts
@@ -7,7 +7,7 @@ export const tool = defineTool({
   path: '/json-to-yaml-converter',
   description: translate('tools.json-to-yaml-converter.description'),
   keywords: ['yaml', 'to', 'json'],
-  component: () => import('./json-to-yaml.vue'),
+  component: () => import('./json-to-yaml-converter.vue'),
   icon: Braces,
   createdAt: new Date('2023-04-10'),
 });
